requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.340
langchain-community==0.0.2
langchain-openai==0.0.2
chromadb==0.4.22
tiktoken==0.5.2
python-dotenv==1.0.0
faiss-cpu==1.12.0
rank-bm25==0.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
nest-asyncio==1.5.8
unstructured==0.11.6
pyngrok==7.0.0
requests==2.32.4
beautifulsoup4==4.12.2
aiohttp==3.9.1
lxml==4.9.3
aiofiles==23.2.1
nltk==3.8.1
pypdf==3.17.0
Pillow>=10.0.0
python-pptx==0.6.23
python-docx==1.1.0
alembic
psycopg2-binary==2.9.9
pysqlite3-binary==0.5.4
playwright==1.49.0
